requirements.txt
pandas
scikit-learn
torch
pillow
torchvision
torchinfo
streamlit
ipykernel
tqdm
matplotlib
visualkeras
-e .